+++ Functions/vyos.ps1
@@ -8,8 +8,8 @@ Function New-SSHSharedKey{
     .DESCRIPTION
     Generate Pre-shared key for SSH authentication to the remote device
 
-    .PARAMETER DestinationIP
-    Mandatory. Must specify destination IP
+    .PARAMETER IP
+    Mandatory. Must specify SSH IP
 
     .PARAMETER User
     Specify user for ssh; defaults to root
@@ -18,7 +18,7 @@ Function New-SSHSharedKey{
     Overwrite current ssh key
 
     .EXAMPLE
-    New-SSHSharedKey -DestinationIP $VyOSExternalIP -User 'vyos' -Force
+    New-SSHSharedKey -IP $VyOSExternalIP -User 'vyos' -Force
 
     .LINK
     http://vcloud-lab.com/entries/devops/How-to-Setup-Passwordless-SSH-Login-on-Windows
@@ -34,7 +34,8 @@ Function New-SSHSharedKey{
 	param
 	(
         [Parameter(Mandatory=$true,Position=0,ValueFromPipeline=$true)]
-		[string]$DestinationIP,
+        [ValidateScript({Test-IPAddress $_})]
+		[string]$IP,
 		[string]$User = 'root',
 		[switch]$Force
 	)
@@ -44,19 +45,24 @@ Function New-SSHSharedKey{
     	Set-Location -Path $env:USERPROFILE
 
         $exeNotExists = @()
-        $CmdAvailable = $true
-        If(!(Test-Command ssh-keygen)){$exeNotExists += 'ssh-keygen';$CmdAvailable = $false}
-        If(!(Test-Command ssh)){$exeNotExists += 'ssh';$CmdAvailable = $false}
-        If(!(Test-Command scp)){$exeNotExists += 'scp';$CmdAvailable = $false}
-    }
-    Process{
-        If($CmdAvailable){
-            Write-Host ("START: Generating Pre-shared key for SSH authentication") -ForegroundColor Green
-            Write-Host ("INFO: You may be prompted a few times to login to {0}..." -f $DestinationIP) -ForegroundColor Gray
-        }Else{
-            Write-Host ("ERROR: {0} does not exist on host, install Git to use SSH" -f ($exeNotExists -join ',')) -ForegroundColor Red
+        If(!(Test-Command ssh-keygen)){$exeNotExists += 'ssh-keygen'}
+        If(!(Test-Command ssh)){$exeNotExists += 'ssh'}
+        If(!(Test-Command scp)){$exeNotExists += 'scp'}
+
+        If($exeNotExists.count -eq 1){
+            Write-Error ("{0} does not exist on host, install Git to use executable" -f $exeNotExists[0])
             return $false
         }
+        ElseIf($exeNotExists.count -gt 1){
+            Write-Error ("{0} do not exist on host, install Git to use executables" -f ($exeNotExists -join ','))
+            return $false
+        }
+        Else{
+            Write-Verbose ("Generating Pre-shared key for SSH authentication")
+            Write-Host ("You may be prompted a few times to login to {0}..." -f $IP) -ForegroundColor Gray
+        }
+    }
+    Process{
 
     	Write-Host "INFO: Checking $env:USERPROFILE/.ssh/id_rsa exists" -ForegroundColor Cyan
     	if (-not(Test-Path -Path "./.ssh/id_rsa") -or $PSBoundParameters.ContainsKey('Force') )
@@ -79,27 +85,27 @@ Function New-SSHSharedKey{
             If($sshrsakey.ExitCode -eq 0){Write-Host "INFO: Generated $env:USERPROFILE\.ssh\id_rsa file" -ForegroundColor Cyan;}
             Else{Return $sshrsakey.ExitCode}
 
-            #TEST Get-Content "./.ssh/known_hosts" | Where-Object {$_ -match $DestinationIP}
+            #TEST Get-Content "./.ssh/known_hosts" | Where-Object {$_ -match $IP}
             If(Test-Path "./.ssh/known_hosts"){
                 $KnownHostContent = Get-Content "./.ssh/known_hosts"
-        		#Get-Content "./.ssh/known_hosts" | Where-Object {$_ -notmatch $DestinationIP} | Set-Content "./.ssh/known_hosts" -Force
+        		#Get-Content "./.ssh/known_hosts" | Where-Object {$_ -notmatch $IP} | Set-Content "./.ssh/known_hosts" -Force
                 $KnownHostContent | ForEach-Object {
-                    $_ | Where-Object {$_ -notmatch $DestinationIP}
+                    $_ | Where-Object {$_ -notmatch $IP}
                 } | Out-File "./.ssh/known_hosts" -Encoding utf8
             }
 
             #TEST Get-content "$env:USERPROFILE/.ssh/id_rsa"
     		$id_rsa_Location = "./.ssh/id_rsa"
-    	    #TEST $remoteSSHServerLogin = "vyos@$DestinationIP"
-    	    $remoteSSHServerLogin = "$User@$DestinationIP"
+    	    #TEST $remoteSSHServerLogin = "vyos@$IP"
+    	    $remoteSSHServerLogin = "$User@$IP"
             If(Test-Path "$env:USERPROFILE\.ssh\id_rsa.pub"){
-                Write-Host "INFO: Copying $env:USERPROFILE\.ssh\id_rsa.pub to $DestinationIP..." -NoNewLine
+                Write-Host "INFO: Copying $env:USERPROFILE\.ssh\id_rsa.pub to $IP..." -NoNewLine
                 start-sleep 5
                 Write-Host "When prompted, please type $User password..." -ForegroundColor Cyan
                 #Start-ExeProcess scp -Arguments "-o StrictHostKeyChecking no '$id_rsa_Location.pub' '${remoteSSHServerLogin}:~/tmp.pub'" -PassThru -Wait
                 Write-Verbose "scp -o 'StrictHostKeyChecking no' `"$id_rsa_Location.pub`" `"${remoteSSHServerLogin}:~/tmp.pub`""
                 scp -o 'StrictHostKeyChecking no' "$id_rsa_Location.pub" "${remoteSSHServerLogin}:~/tmp.pub"
-                Write-Host "INFO: Updating authorized_keys on $DestinationIP..." -NoNewLine
+                Write-Host "INFO: Updating authorized_keys on $IP..." -NoNewLine
                 start-sleep 5
                 Write-Verbose "ssh `"${remoteSSHServerLogin}`" `"mkdir -p ~/.ssh && chmod 700 ~/.ssh && cat ~/tmp.pub >> ~/.ssh/authorized_keys && chmod 600 ~/.ssh/authorized_keys && rm -f ~/tmp.pub`""
                 ssh "${remoteSSHServerLogin}" "mkdir -p ~/.ssh && chmod 700 ~/.ssh && cat ~/tmp.pub >> ~/.ssh/authorized_keys && chmod 600 ~/.ssh/authorized_keys && rm -f ~/tmp.pub"
@@ -114,7 +120,7 @@ Function New-SSHSharedKey{
             If($sshrsakey.ExitCode -ne 0){Return $sshrsakey.ExitCode}
     	}
 
-    	#Write-Host "DONE: Try running command: 'ssh $User@$DestinationIP'; Now it will not prompt for password" -ForegroundColor Green
+    	#Write-Host "DONE: Try running command: 'ssh $User@$IP'; Now it will not prompt for password" -ForegroundColor Green
     }
     End{
         Set-Location -Path $oldLocation
@@ -141,6 +147,12 @@ function ConvertTo-LinuxLineEndings($path) {
 }
 
 Function New-VyattaScript{
+    [CmdletBinding(
+		SupportsShouldProcess=$True,
+		ConfirmImpact='Medium',
+		HelpURI='http://vcloud-lab.com',
+		DefaultParameterSetName='Manual'
+	)]
     Param (
         $Value,
         [string]$ExportPath,
@@ -175,30 +187,39 @@ Function New-VyattaScript{
     }
 }
 
-Function Initialize-VyattaScript {
-    [CmdletBinding()]
+Function Invoke-VyattaScript {
+    [CmdletBinding(
+		SupportsShouldProcess=$True,
+		ConfirmImpact='Medium',
+		HelpURI='http://vcloud-lab.com',
+		DefaultParameterSetName='Manual'
+	)]
     Param (
         [Parameter(Mandatory=$true,Position=0)]
         [ValidateScript({Test-IPAddress $_})]
         [string]$IP,
         [string]$RSAFile = "$env:USERPROFILE\.ssh\id_rsa",
         [Parameter(Mandatory=$true,Position=1)]
-        [string]$Path,
-        [switch]$Execute
+        [string]$Path
     )
 
     #copy script to vyos router
     $remoteSSHServerLogin = "vyos@$IP"
 
     $exeNotExists = @()
-    $CmdAvailable = $true
-    If(!(Test-Command ssh-keygen)){$exeNotExists += 'ssh-keygen';$CmdAvailable = $false}
-    If(!(Test-Command ssh)){$exeNotExists += 'ssh';$CmdAvailable = $false}
-    If(!(Test-Command scp)){$exeNotExists += 'scp';$CmdAvailable = $false}
+    If(!(Test-Command ssh)){$exeNotExists += 'ssh'}
+    If(!(Test-Command scp)){$exeNotExists += 'scp'}
 
-    If(!$CmdAvailable){
-        Write-Host ("ERROR: {0} does not exist on host, install Git to use SSH" -f ($exeNotExists -join ',')) -ForegroundColor Red
+    If($exeNotExists.count -eq 1){
+        Write-Error ("{0} does not exist on host, install Git to use executable" -f $exeNotExists[0])
         return $false
+    }
+    ElseIf($exeNotExists.count -gt 1){
+        Write-Error ("{0} do not exist on host, install Git to use executables" -f ($exeNotExists -join ','))
+        return $false
+    }
+    Else{
+        Write-Verbose ("Executables exist on host. Attempting to run script...")
     }
 
     If((Test-Path $Path) -and (Test-Connection $IP -Count 1))
@@ -211,38 +232,85 @@ Function Initialize-VyattaScript {
         }
 
         #TEST scp -o 'StrictHostKeyChecking no' "${remoteSSHServerLogin}:~/.scripts/test.sh" "$env:temp\test.bh"
-
-        If($PSBoundParameters.ContainsKey('Execute'))
-        {
-            $randomchar = -join ((65..90) + (97..122) | Get-Random -Count 5 | % {[char]$_})
-            $scriptfile = "vyos_" + $randomchar.ToLower() + ".sh"
-            Write-Verbose "Executing file 'tmp.sh' as $scriptfile on $IP"
-            #build bash command
-            $bashCommands = @(
-                'mkdir -p ~/.scripts'
-                'chmod 700 ~/.scripts'
-                "rm -f ~/.scripts/$scriptfile"
-                "cat ~/tmp.sh >> ~/.scripts/$scriptfile"
-                'rm -f ~/tmp.sh'
-                "sed -i -e 's/\r$//' ~/.scripts/$scriptfile"
-                "chmod u+x ~/.scripts/$scriptfile"
-                "sg vyattacfg -c ~/.scripts/$scriptfile"
-            )
-            #join all commands as single line separated with &&
-            $bashCommand = $bashCommands -join ' && '
-            If(Test-Path $RSAFile){
-                Write-Verbose "ssh -i `"$RSAFile`" `"$remoteSSHServerLogin`" $bashCommand"
-                #VERBOSE & ssh -v -i $RSAFile $remoteSSHServerLogin $bashCommand
-                ssh -i "$RSAFile" "$remoteSSHServerLogin" $bashCommand
-            }Else{
-                Write-Verbose "ssh `"$remoteSSHServerLogin`" $bashCommand"
-                ssh "$remoteSSHServerLogin" $bashCommand
-            }
-
+        $randomchar = -join ((65..90) + (97..122) | Get-Random -Count 5 | % {[char]$_})
+        $scriptfile = "vyos_" + $randomchar.ToLower() + ".sh"
+        Write-Verbose "Executing file 'tmp.sh' as $scriptfile on $IP"
+        #build bash command
+        $bashCommands = @(
+            'mkdir -p ~/.scripts'
+            'chmod 700 ~/.scripts'
+            "rm -f ~/.scripts/$scriptfile"
+            "cat ~/tmp.sh >> ~/.scripts/$scriptfile"
+            'rm -f ~/tmp.sh'
+            "sed -i -e 's/\r$//' ~/.scripts/$scriptfile"
+            "chmod u+x ~/.scripts/$scriptfile"
+            "sg vyattacfg -c ~/.scripts/$scriptfile"
+        )
+        #join all commands as single line separated with &&
+        $bashCommand = $bashCommands -join ' && '
+        If(Test-Path $RSAFile){
+            Write-Verbose "ssh -i `"$RSAFile`" `"$remoteSSHServerLogin`" $bashCommand"
+            #VERBOSE & ssh -v -i $RSAFile $remoteSSHServerLogin $bashCommand
+            ssh -i "$RSAFile" "$remoteSSHServerLogin" $bashCommand
+        }Else{
+            Write-Verbose "ssh `"$remoteSSHServerLogin`" $bashCommand"
+            ssh "$remoteSSHServerLogin" $bashCommand
         }
     }
     Else{
         Throw "Unable to find file: $Path"
+    }
+
+}
+
+Function Invoke-VyattaCmd {
+    [CmdletBinding(
+		SupportsShouldProcess=$True,
+		ConfirmImpact='Medium',
+		HelpURI='http://vcloud-lab.com',
+		DefaultParameterSetName='Manual'
+	)]
+    Param (
+        [Parameter(Mandatory=$true,Position=0)]
+        [ValidateScript({Test-IPAddress $_})]
+        [string]$IP,
+        [string]$RSAFile = "$env:USERPROFILE\.ssh\id_rsa",
+        [Parameter(Mandatory=$true,Position=1,ValueFromPipeline=$true)]
+        [string]$Cmd
+    )
+    Begin{
+        #copy script to vyos router
+        $remoteSSHServerLogin = "vyos@$IP"
+
+        $exeNotExists = @()
+        If(!(Test-Command ssh)){$exeNotExists += 'ssh'}
+
+        If($exeNotExists.count -eq 1){
+            Write-Error ("{0} does not exist on host, install Git to use executable" -f $exeNotExists[0])
+            return $false
+        }
+        ElseIf($exeNotExists.count -gt 1){
+            Write-Error ("{0} do not exist on host, install Git to use executables" -f ($exeNotExists -join ','))
+            return $false
+        }
+        Else{
+            Write-Verbose ("Executables exist on host. Attempting to run command...")
+        }
+
+    }
+    Process{
+
+        If(Test-Path $RSAFile){
+            Write-Verbose "RUNNING COMMAND: ssh `"${remoteSSHServerLogin}`" `"$Cmd`""
+            ssh "${remoteSSHServerLogin}" "$Cmd"
+        }
+        Else{
+            Write-Verbose "RUNNING COMMAND: ssh -i `"$RSAFile`" `"${remoteSSHServerLogin}`" `"$Cmd`""
+            ssh -i "$RSAFile" "${remoteSSHServerLogin}" "$Cmd"
+        }
+    }
+    End{
+
     }
 
 }