+++ Step 3A. Build Azure Basic S2S.ps1
@@ -372,7 +372,7 @@ save
 #region 11: Build reset vpn config
 $VyOSReset = @"
 restart vpn
-show ipsec vpn sa
+run show ipsec vpn sa
 `n
 "@
 #endregion
@@ -390,9 +390,9 @@ If($RouterAutomationMode)
     $VyOSFinalScript = New-VyattaScript -Value $VyOSFinal -AsObject -SetReboot
     #TEST $VyOSFinalScript.value
     #temporary set auto logon ssh keys
-    New-SSHSharedKey -DestinationIP $VyOSExternalIP -User 'vyos' -Verbose
+    New-SSHSharedKey -IP $VyOSExternalIP -User 'vyos' -Verbose
 
-    $Result = Initialize-VyattaScript -IP $VyOSExternalIP -Path $VyOSFinalScript.Path -Execute -Verbose
+    $Result = Invoke-VyattaScript -IP $VyOSExternalIP -Path $VyOSFinalScript.Path -Verbose
 
     $Result
 
@@ -461,9 +461,9 @@ public class TrustAllCertsPolicy : ICertificatePolicy {
                 Start-Sleep 10
                 $VyOSResetScript = New-VyattaScript -Value $VyOSReset -AsObject -SetReboot
                 #TEST $VyOSFinalScript.value
-                New-SSHSharedKey -DestinationIP $VyOSExternalIP -User 'vyos' -Verbose
+                New-SSHSharedKey -IP $VyOSExternalIP -User 'vyos' -Verbose
 
-                $Result = Initialize-VyattaScript -IP $VyOSExternalIP -Path $VyOSResetScript.Path -Execute -Verbose
+                $Result = Invoke-VyattaScript -IP $VyOSExternalIP -Path $VyOSResetScript.Path -Verbose
                 If(!$Result){
                     Write-Host "Failed to reset the vpn on vyos router; use manual process" -ForegroundColor Red
                 }